tests/wpt/tests/html/semantics/scripting-1/the-script-element/execution-timing/099.html

<!DOCTYPE html>
<html><head>
        <title> scheduler: defer adding iframe containing script</title>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="testlib/testlib.js"></script>
</head>
<body>

        <div id="log">FAILED (This TC requires JavaScript enabled)</div>

        <script defer src="scripts/include-11.js"></script>
        <script>
          var t = async_test();

          onload = t.step_func(function() {assert_array_equals(eventOrder, ["external script before adding iframe", "script in iframe"]); t.done();});
        </script>
</body>
</html>
